VerificarCadastro method moved from AlunoController to AlunoDAO
