+++ sistemaBiblioteca/src/utils/ConnectionSQL.java
@@ -7,7 +7,7 @@ import java.sql.SQLException;
 public class ConnectionSQL {
     private static final String URL = "jdbc:mysql://localhost:3306/sistemabiblioteca";
     private static final String USER = "root"; 
-    private static final String PASSWORD = "root"; 
+    private static final String PASSWORD = "admin"; 
 
     public static Connection getConnection() {
         try {
